Version 0.50 Release
Removed 5,000 sample averaging in the IOIO code and replaced it with 1
second sampling and an average of the past 30 seconds in
FurnaceController.

Added init function to webserver_js to properly set which temperature
drop downs are visible.

Updated ping out code to verify a valid url is set.

Updated the settings json post code to properly check if the location
has changed.

Updated the conditions timer to 1 second instead of 5 seconds.

Updated Furnace controller to turn on the fan when turning on the AC.
This is necessary in some systems where the AC is controlled separately
from the fan.